halfway terminated
This project is terminated and will not be maintained any more. The Oauth2 authentication function doesn't work(many files are not completed).

+++ src/main/java/com/learn/bulletin/entity/News.java
@@ -12,10 +12,10 @@ import java.util.List;
 public class News {
 
     @ApiModelProperty(value = "所属用户id",example = "0")
-    private Integer user_id;
+    private Long user_id;
 
     @ApiModelProperty(hidden = true ,example = "0")
-    private Integer news_id;
+    private Long news_id;
 
     @ApiModelProperty(value = "新闻标题", example = "Title")
     private String title;
